move apiserver options
