+++ federation/cmd/federation-apiserver/app/options/BUILD
@@ -16,9 +16,9 @@ go_library(
     tags = ["automanaged"],
     deps = [
         "//pkg/features:go_default_library",
-        "//pkg/genericapiserver/server/options:go_default_library",
         "//pkg/kubeapiserver/options:go_default_library",
         "//vendor:github.com/spf13/pflag",
+        "//vendor:k8s.io/apiserver/pkg/server/options",
     ],
 )
