+++ pkg/genericapiserver/server/BUILD
@@ -25,7 +25,6 @@ go_test(
         "//pkg/client/clientset_generated/clientset:go_default_library",
         "//pkg/generated/openapi:go_default_library",
         "//pkg/genericapiserver/registry/rest:go_default_library",
-        "//pkg/genericapiserver/server/options:go_default_library",
         "//pkg/storage/etcd/testing:go_default_library",
         "//vendor:github.com/go-openapi/spec",
         "//vendor:github.com/stretchr/testify/assert",
@@ -41,6 +40,7 @@ go_test(
         "//vendor:k8s.io/apiserver/pkg/authentication/user",
         "//vendor:k8s.io/apiserver/pkg/authorization/authorizer",
         "//vendor:k8s.io/apiserver/pkg/endpoints/request",
+        "//vendor:k8s.io/apiserver/pkg/server/options",
         "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
         "//vendor:k8s.io/apiserver/pkg/util/flag",
         "//vendor:k8s.io/client-go/util/cert",
@@ -70,7 +70,6 @@ go_library(
         "//pkg/genericapiserver/endpoints/openapi:go_default_library",
         "//pkg/genericapiserver/registry/rest:go_default_library",
         "//pkg/genericapiserver/server/mux:go_default_library",
-        "//pkg/genericapiserver/server/options:go_default_library",
         "//pkg/genericapiserver/server/routes:go_default_library",
         "//vendor:github.com/coreos/go-systemd/daemon",
         "//vendor:github.com/emicklei/go-restful",
@@ -102,6 +101,7 @@ go_library(
         "//vendor:k8s.io/apiserver/pkg/endpoints/request",
         "//vendor:k8s.io/apiserver/pkg/server/filters",
         "//vendor:k8s.io/apiserver/pkg/server/healthz",
+        "//vendor:k8s.io/apiserver/pkg/server/options",
         "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
         "//vendor:k8s.io/client-go/rest",
         "//vendor:k8s.io/client-go/util/cert",
@@ -121,7 +121,6 @@ filegroup(
         ":package-srcs",
         "//pkg/genericapiserver/server/mux:all-srcs",
         "//pkg/genericapiserver/server/openapi:all-srcs",
-        "//pkg/genericapiserver/server/options:all-srcs",
         "//pkg/genericapiserver/server/routes:all-srcs",
     ],
     tags = ["automanaged"],
