+++ pkg/kubeapiserver/options/BUILD
@@ -23,12 +23,12 @@ go_library(
         "//pkg/cloudprovider:go_default_library",
         "//pkg/controller/informers:go_default_library",
         "//pkg/genericapiserver/server:go_default_library",
-        "//pkg/genericapiserver/server/options:go_default_library",
         "//pkg/kubeapiserver/authenticator:go_default_library",
         "//pkg/kubeapiserver/authorizer:go_default_library",
         "//vendor:github.com/golang/glog",
         "//vendor:github.com/spf13/pflag",
         "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
+        "//vendor:k8s.io/apiserver/pkg/server/options",
     ],
 )
