+++ staging/src/k8s.io/apiserver/pkg/server/options/doc.go
@@ -18,4 +18,4 @@ limitations under the License.
 // server. It takes a minimal set of dependencies and does not reference
 // implementations, in order to ensure it may be reused by multiple components
 // (such as CLI commands that wish to generate or validate config).
-package options // import "k8s.io/kubernetes/pkg/genericapiserver/server/options"
+package options // import "k8s.io/apiserver/pkg/server/options"
